Fixed comp routes

+++ routes/api/compRoute.js
@@ -7,7 +7,7 @@ const verifyToken = require("./verifyToken");
 module.exports = (db) => {
    // Select Comps
    router.get("/", verifyToken, (req, res) => {
-      let sql = "SELECT * FROM competitions";
+      let sql = `SELECT * FROM competitions WHERE Username = '${req.auth.user.username}'`;
       db.query(sql, (err, rows) => {
          if (err) {
             res.status(404).send(err);
@@ -18,8 +18,8 @@ module.exports = (db) => {
 
    // Insert Comp
    router.post("/", verifyToken, (req, res) => {
-      let sql = `INSERT INTO competitions (CompetitionName) VALUES ('${req.body.competitionName}')`;
-      db.query(sql, (err, rows) => {
+      let sql = `INSERT INTO competitions (Username, CompetitionName) VALUES ('${req.auth.user.username}', '${req.body.competitionName}')`;
+      db.query(sql, (err, result) => {
          if (err) {
             res.status(404).send(err);
          }
@@ -29,18 +29,19 @@ module.exports = (db) => {
 
    // Update Comp
    router.patch("/:id", verifyToken, (req, res) => {
-      let sql = `UPDATE competitions SET CompetitionName = '${req.body.competitionName}' WHERE ID = '${req.params.id}'`;
-      db.query(sql, (err) => {
+      let sql = `UPDATE competitions SET CompetitionName = '${req.body.competitionName}' WHERE ID = '${req.params.id}' AND Username = '${req.auth.user.username}'`;
+      db.query(sql, (err, result) => {
          if (err) {
             res.status(404).send(err);
          }
-         res.send();
+         if (result.affectedRows) res.send();
+         else res.status(204).send();
       });
    });
 
    // Delete Comp
    router.delete("/:id", verifyToken, (req, res) => {
-      let sql = `DELETE FROM competitions WHERE ID = '${req.params.id}' LIMIT 1`;
+      let sql = `DELETE FROM competitions WHERE ID = '${req.params.id}' AND Username = '${req.auth.user.username}' LIMIT 1`;
       db.query(sql, (err, result) => {
          if (err) {
             res.status(404).send(err);
